# app/api/v1/docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "MercadoLibre Challenge API",
    "description": "API para el challenge de la página de detalle de producto de MercadoLibre",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "API Base URL"
    }
  ],
  "paths": {
...
    "/products": {
      "get": {
        "summary": "Obtener listado de productos",
        "description": "Obtiene un listado paginado de productos con filtros opcionales",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1
            },
            "description": "Número de página"
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 10
            },
            "description": "Número de productos por página"
          },
          {
            "name": "storeId",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "ID de la tienda para filtrar productos"
          },
          {
            "name": "secondaryCategoryId",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "ID de categoría secundaria para filtrar productos"
          },
          {
            "name": "excludeId",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "ID de producto a excluir de los resultados"
          }
        ],
        "responses": {
          "200": {
            "description": "Listado paginado de productos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductsResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ProductPreview": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "ID del producto"
          },
          "title": {
            "type": "string",
            "description": "Título del producto"
          },
          "storeData": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "description": "ID de la tienda"
              },
              "name": {
                "type": "string",
                "description": "Nombre de la tienda"
              }
            }
          },
          "mainImage": {
            "type": "string",
            "format": "byte",
            "description": "Imagen principal del producto (base64)"
          },
          "priceConfig": {
            "type": "object",
            "properties": {
              "price": {
                "type": "number",
                "format": "float",
                "description": "Precio del producto"
              },
              "currency": {
                "type": "string",
                "enum": ["USD", "ARS"],
                "description": "Moneda del precio"
              }
            }
          },
          "isBestSelling": {
            "type": "boolean",
            "description": "Indica si el producto es más vendido"
          },
          "isNew": {
            "type": "boolean",
            "description": "Indica si el producto es nuevo"
          }
        }
      },
      "ProductsResponse": {
        "type": "object",
        "properties": {
          "page": {
            "type": "integer",
            "description": "Número de página actual"
          },
          "limit": {
            "type": "integer",
            "description": "Número de productos por página"
          },
          "total": {
            "type": "integer",
            "description": "Total de productos"
          },
          "totalPages": {
            "type": "integer",
            "description": "Total de páginas"
          },
          "products": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProductPreview"
            }
          }
        }
      }
    }
  }
}